DPC-30: Refactored Auth test into helper class
Moved some of test utils into a helper class. This will be more useful for the work in DPC-308.
